Eluna changes
* Attempt fix VS 2010 pow error
* Tweak the doc parser a little to work with variable arugments ... and show
enum names and similar.
* Change push and check to work with more types like size and time (fix osx)
* Add errors to using not implemented operations on eluna objects
* Change some SQL comments to work with documentation better
* Add functions to create uint64 and int64 values (may need rename)
* Change doc generation documentation a little to make it clearer how proto is used

+++ CorpseMethods.h
@@ -31,7 +31,7 @@ namespace LuaCorpse
      */
     int GetGhostTime(Eluna* /*E*/, lua_State* L, Corpse* corpse)
     {
-        Eluna::Push(L, uint32(corpse->GetGhostTime()));
+        Eluna::Push(L, corpse->GetGhostTime());
         return 1;
     }
